🚚 Move validation to client side
