+++ src/routes/ngs-api/runs/duels/indomitable/[boss=indomitableBoss]/+server.ts
@@ -5,7 +5,7 @@ import { parseIndomitableBoss } from '$lib/types/api/duels/indomitableBoss.js';
 import {
 	indomitableRunSearchFilterSchema,
 	type IndomitableRunSearchFilter
-} from '$lib/server/types/validation/indomitableRunFilter.js';
+} from '$lib/types/api/validation/indomitableRunFilter.js';
 import { getIndomitableRuns } from '$lib/server/repositories/indomitableRunsRepository.js';
 import { mapIndomitableToRun } from '$lib/server/mappers/api/indomitableRunMapper.js';
 import { parseToRawSchema } from '$lib/utils/schemaValidation.js';
