+++ src/routes/ngs-api/submissions/[category=categoryDfa]/+server.ts
@@ -19,7 +19,7 @@ import { jsonError } from '$lib/server/error.js';
 import {
 	dfaSubmissionRequestSchema,
 	type DfaSubmissionRequest
-} from '$lib/server/types/validation/dfaSubmissions.js';
+} from '$lib/types/api/validation/dfaSubmissions.js';
 import { weaponsToDbValMap } from '$lib/server/db/util/weaponType.js';
 import { parseWeapon } from '$lib/types/api/weapon.js';
 import { normalizeYoutubeLink } from '$lib/utils/youtube.js';
