+++ src/routes/ngs-api/submissions/[category=categoryPurples]/approve/+server.ts
@@ -6,7 +6,7 @@ import { notifyDiscordNewRunApproved } from '$lib/server/discordNotify.js';
 import {
 	approveRequestSchema,
 	type ApproveRequest
-} from '$lib/server/types/validation/submissions.js';
+} from '$lib/types/api/validation/submissions.js';
 import { getRunPlayer } from '$lib/server/repositories/playerRepository.js';
 import { RunCategories, parseRunCategory } from '$lib/types/api/categories.js';
 import {
